feat: Upgrade to .NET Standard 2.0 and reorg folder structure
Thus far this concentrates only on the actual library. There are actual code changes that are needed to clean up the warnings, but this at least organizes the folders in a cleaner architecture.

Converting the demo and texturetest projects is a larger project.

"demo" is what was the "renderXF" GUI application, split off from the renderxf library. There's a bit of a source issue here as the form accesses the internals during shutdown.